pyproject.toml
[project]
name = "llm-discord-bot"
version = "0.2.0"
description = "LLM based discord bot. Optimised for entartained and mimicking human responses."
requires-python = ">=3.12"

dependencies = [
    "discord.py>=2.3.0",
    "httpx>=0.27.0",
    "ollama>=0.4.0",
    "pyyaml>=6.0",
]
